refactor(example): split metal_overlay into modules
