+++ examples/metal_overlay/overlay.rs
@@ -1,0 +1,236 @@
+//! Menu state and configuration
+
+use screencapturekit::prelude::*;
+
+pub struct OverlayState {
+    pub show_help: bool,
+    pub show_waveform: bool,
+    pub show_config: bool,
+    pub config_selection: usize,
+    pub menu_selection: usize,
+}
+
+impl OverlayState {
+    pub fn new() -> Self {
+        Self {
+            show_help: true,
+            show_waveform: true,
+            show_config: false,
+            config_selection: 0,
+            menu_selection: 0,
+        }
+    }
+
+    pub const MENU_ITEMS: &'static [&'static str] = &["Picker", "Capture", "Config", "Quit"];
+
+    pub fn menu_count() -> usize {
+        Self::MENU_ITEMS.len()
+    }
+}
+
+pub struct ConfigMenu;
+
+impl ConfigMenu {
+    pub const OPTIONS: &'static [&'static str] = &[
+        "FPS",
+        "Cursor",
+        "Audio",
+        "Mic",
+        "Mic Device",
+        "Exclude Self",
+        "Sample Rate",
+        "Channels",
+        "Scale Fit",
+        "Aspect",
+        "Retina",
+        "Opaque",
+        "Shadows Only",
+        "Ignore Shadows",
+        "Pixel Format",
+        "Queue",
+    ];
+    pub const FPS_OPTIONS: [u32; 5] = [15, 30, 60, 120, 240];
+    pub const QUEUE_OPTIONS: [u32; 4] = [3, 5, 8, 12];
+    pub const SAMPLE_RATE_OPTIONS: [i32; 3] = [44100, 48000, 96000];
+    pub const CHANNEL_OPTIONS: [i32; 2] = [1, 2];
+
+    pub fn option_count() -> usize {
+        Self::OPTIONS.len()
+    }
+
+    pub fn option_name(idx: usize) -> &'static str {
+        Self::OPTIONS.get(idx).unwrap_or(&"?")
+    }
+
+    pub fn option_value(
+        config: &SCStreamConfiguration,
+        mic_device_idx: Option<usize>,
+        idx: usize,
+    ) -> String {
+        match idx {
+            0 => format!("{}", config.fps()),
+            1 => if config.shows_cursor() { "On" } else { "Off" }.to_string(),
+            2 => if config.captures_audio() { "On" } else { "Off" }.to_string(),
+            3 => if config.captures_microphone() { "On" } else { "Off" }.to_string(),
+            4 => match mic_device_idx {
+                None => "Default".to_string(),
+                Some(idx) => AudioInputDevice::list()
+                    .get(idx)
+                    .map(|d| d.name.chars().take(10).collect())
+                    .unwrap_or_else(|| "?".to_string()),
+            },
+            5 => if config.excludes_current_process_audio() { "On" } else { "Off" }.to_string(),
+            6 => format!("{}Hz", config.sample_rate()),
+            7 => format!("{}ch", config.channel_count()),
+            8 => if config.scales_to_fit() { "On" } else { "Off" }.to_string(),
+            9 => if config.preserves_aspect_ratio() { "On" } else { "Off" }.to_string(),
+            10 => if config.increase_resolution_for_retina_displays() { "On" } else { "Off" }.to_string(),
+            11 => if config.should_be_opaque() { "On" } else { "Off" }.to_string(),
+            12 => if config.captures_shadows_only() { "On" } else { "Off" }.to_string(),
+            13 => if config.ignores_shadows_display() { "On" } else { "Off" }.to_string(),
+            14 => format!("{:?}", config.pixel_format()),
+            15 => format!("{}", config.queue_depth()),
+            _ => "?".to_string(),
+        }
+    }
+
+    pub fn toggle_or_adjust(
+        config: &mut SCStreamConfiguration,
+        mic_device_idx: &mut Option<usize>,
+        idx: usize,
+        increase: bool,
+    ) {
+        use screencapturekit::stream::configuration::pixel_format::PixelFormat;
+        const PIXEL_FORMATS: [PixelFormat; 4] = [
+            PixelFormat::BGRA,
+            PixelFormat::l10r,
+            PixelFormat::YCbCr_420v,
+            PixelFormat::YCbCr_420f,
+        ];
+
+        match idx {
+            0 => {
+                // FPS
+                let current_fps = config.fps();
+                let current_idx = Self::FPS_OPTIONS.iter().position(|&f| f == current_fps).unwrap_or(2);
+                let new_idx = if increase {
+                    (current_idx + 1) % Self::FPS_OPTIONS.len()
+                } else {
+                    (current_idx + Self::FPS_OPTIONS.len() - 1) % Self::FPS_OPTIONS.len()
+                };
+                config.set_fps(Self::FPS_OPTIONS[new_idx]);
+            }
+            1 => { config.set_shows_cursor(!config.shows_cursor()); }
+            2 => { config.set_captures_audio(!config.captures_audio()); }
+            3 => { config.set_captures_microphone(!config.captures_microphone()); }
+            4 => {
+                // Mic Device
+                let devices = AudioInputDevice::list();
+                if devices.is_empty() {
+                    return;
+                }
+                match *mic_device_idx {
+                    None => {
+                        *mic_device_idx = Some(if increase { 0 } else { devices.len() - 1 })
+                    }
+                    Some(idx) => {
+                        if increase {
+                            *mic_device_idx = if idx + 1 >= devices.len() {
+                                None
+                            } else {
+                                Some(idx + 1)
+                            };
+                        } else {
+                            *mic_device_idx = if idx == 0 { None } else { Some(idx - 1) };
+                        }
+                    }
+                }
+                if let Some(idx) = *mic_device_idx {
+                    if let Some(device) = devices.get(idx) {
+                        config.set_microphone_capture_device_id(&device.id);
+                    }
+                }
+            }
+            5 => { config.set_excludes_current_process_audio(!config.excludes_current_process_audio()); }
+            6 => {
+                // Sample Rate
+                let current = config.sample_rate();
+                let current_idx = Self::SAMPLE_RATE_OPTIONS
+                    .iter()
+                    .position(|&r| r == current)
+                    .unwrap_or(1);
+                let new_idx = if increase {
+                    (current_idx + 1) % Self::SAMPLE_RATE_OPTIONS.len()
+                } else {
+                    (current_idx + Self::SAMPLE_RATE_OPTIONS.len() - 1) % Self::SAMPLE_RATE_OPTIONS.len()
+                };
+                config.set_sample_rate(Self::SAMPLE_RATE_OPTIONS[new_idx]);
+            }
+            7 => {
+                // Channels
+                let current = config.channel_count();
+                let current_idx = Self::CHANNEL_OPTIONS
+                    .iter()
+                    .position(|&c| c == current)
+                    .unwrap_or(1);
+                let new_idx = if increase {
+                    (current_idx + 1) % Self::CHANNEL_OPTIONS.len()
+                } else {
+                    (current_idx + Self::CHANNEL_OPTIONS.len() - 1) % Self::CHANNEL_OPTIONS.len()
+                };
+                config.set_channel_count(Self::CHANNEL_OPTIONS[new_idx]);
+            }
+            8 => { config.set_scales_to_fit(!config.scales_to_fit()); }
+            9 => { config.set_preserves_aspect_ratio(!config.preserves_aspect_ratio()); }
+            10 => { config.set_increase_resolution_for_retina_displays(
+                !config.increase_resolution_for_retina_displays(),
+            ); }
+            11 => { config.set_should_be_opaque(!config.should_be_opaque()); }
+            12 => { config.set_captures_shadows_only(!config.captures_shadows_only()); }
+            13 => { config.set_ignores_shadows_display(!config.ignores_shadows_display()); }
+            14 => {
+                // Pixel Format
+                let current = config.pixel_format();
+                let current_idx = PIXEL_FORMATS.iter().position(|&f| f == current).unwrap_or(0);
+                let new_idx = if increase {
+                    (current_idx + 1) % PIXEL_FORMATS.len()
+                } else {
+                    (current_idx + PIXEL_FORMATS.len() - 1) % PIXEL_FORMATS.len()
+                };
+                config.set_pixel_format(PIXEL_FORMATS[new_idx]);
+            }
+            15 => {
+                // Queue
+                let current_depth = config.queue_depth();
+                let current_idx = Self::QUEUE_OPTIONS
+                    .iter()
+                    .position(|&q| q == current_depth)
+                    .unwrap_or(2);
+                let new_idx = if increase {
+                    (current_idx + 1) % Self::QUEUE_OPTIONS.len()
+                } else {
+                    (current_idx + Self::QUEUE_OPTIONS.len() - 1) % Self::QUEUE_OPTIONS.len()
+                };
+                config.set_queue_depth(Self::QUEUE_OPTIONS[new_idx]);
+            }
+            _ => {}
+        }
+    }
+}
+
+pub fn default_stream_config() -> SCStreamConfiguration {
+    SCStreamConfiguration::new()
+        .with_width(1920)
+        .with_height(1080)
+        .with_fps(60)
+        .with_shows_cursor(true)
+        .with_captures_audio(true)
+        .with_excludes_current_process_audio(true)
+        .with_captures_microphone(true)
+        .with_channel_count(2)
+        .with_sample_rate(48000)
+        .with_scales_to_fit(true)
+        .with_preserves_aspect_ratio(true)
+        .with_queue_depth(8)
+        .with_pixel_format(screencapturekit::stream::configuration::PixelFormat::BGRA)
+}
